+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - controller: add an S3 storage interface for saving assets in a long-term storage ([#214](https://git.heinlein-video.de/heinlein-video/k3k-controller/-/issues/214))
 - whiteboard: save generated PDF files in S3 storage ([#225](https://git.heinlein-video.de/heinlein-video/k3k-controller/-/issues/225))
 
+### Removed
+
+- controller: `status` field from event resource ([#221](https://git.heinlein-video.de/heinlein-video/k3k-controller/-/issues/221))
+
 ## [1.0.0-rc.8] - 2022-11-10
 
 ### Added